Move map to component, add to project browse
