+++ app/assets/scripts/components/map.js
